Update Ca1_Programming.java
You have been given a file named “people.txt” – this contains the details of a person in the following
format:
Line 1 --- <firstname> <surname>
Line 2 --- <age>
Line 3 --- <gender>
The firstname and surname should be text; the age should be a number between 0 and 100, and the
gender should be either ‘M’ (for Male), ‘F’ (for Female), or ‘T’ (for Transgender)
Example:
Tiago Murphy
23
M
Your task is to:
1) Read the information and store it in appropriate variables in your program.
2) Check that the data obeys the validation rules. These are:
a. The firstname and surname must be TEXT ONLY. No numbers are allowed. There
must be 1 space in the line to separate firstname from surname. If there is more
than 1 space, everything after the 1st space is considered to be the surname.
b. The age must be an integer between 0 and 100
c. The gender must be only the letter ‘M’, or the letter ‘F’ or the letter ‘T’.
If any of the validation rules have been broken, then you must output an appropriate error
message that tells the user clearly why it is not valid.
3) Output the data to a new file called “status.txt” in the following format:
<title> <surname>, <first initial>
<status>
MLO 1 - Understand and employ fundamental concepts and principles of programming such as
variables, Boolean expressions, control flow structures, methods, arrays, etc.
MLO 2 - demonstrate a structured approach to algorithmic design and problem solving and
exhibit professional development best practices in designing and developing robust,
maintainable software
CA 1 – File Access & String Manipulation Page 3 of 4
The surname must be exactly as input.
The first initial must be the first letter of the firstname that was input. It must be output
after the surname. Note also that a comma must be output between the surname and the
first initial.
The title must be “Mr” if the gender was “M”, or “Ms” if the gender was “F”, or “Mx” if the
gender was “T”
The status must be “School” if the age was 18 or less, or “College” if the age was between 19
and 25, or “Worker” if the age was between 26 and 66, or “Retired” if the age is 67 or over.
DISTINCTION WORK
To gain a distinction (70% or higher) you must complete tasks 1) – 3) properly AND ALSO:
Amend the programme to include a loop that will process ALL the data in the file. You can assume
that the data for the 2nd person begins on the next line immediately after the 1st person.
Remember that you need to check each time whether or not the data is valid. Also remember that
when you output, you need to add each person to the end of the status.txt file.